// helpers/types.ts
export type tEthereumAddress = string;
// ...
export enum eContractid {
    Example = 'Example',
    Unitroller = 'Unitroller',
    Comptroller = 'Comptroller',
    JumpRateModel = 'JumpRateModel',
    PErc20Delegate = 'PErc20Delegate',
    PERC20Delegator = 'PERC20Delegator',
    CEther = 'CEther',
    SimplePriceOracle = 'SimplePriceOracle',
    SimplePriceOracleChainlink = 'SimplePriceOracleChainlink',
    NewAssetsToken = 'NewAssetsToken',
    PubMiningRateModel = 'PubMiningRateModel',
    MarginSwapPool = 'MarginSwapPool',
    PriceOracleAggregator = 'PriceOracleAggregator',
    TestPriceOracle = 'TestPriceOracle', //moac
    // Config = 'Config',
    Controller = 'Controller',
    MockTXAggregator = 'MockTXAggregator',
    MSPStorage = 'MSPStorage',
    Capital='Capital',
    Liquidation='Liquidation',
    PUB='PUB',
    AToken = 'AToken',
}